# clearai-backend/openapi.yaml
openapi: 3.0.3
info:
  title: ClearAI Backend API
  version: 0.2.0
  description: |
    HTTP surface of the ClearAI customs-classification backend. Source of
    truth for the Fastify route handlers in `clearai-backend/src/`. Imported
    by Azure API Management to define the public API shape.

    2026-05-12 rename cutover (commit 06c9f3d, backend rev 70):
      - /declaration-runs/*           -> /batches/*
      - /declaration-runs/:id/classifications -> /batches/:id/items
      - PATCH /declaration-runs/:id    -> POST /batches/:id/cancel
      - /declaration-runs/:id/download-links -> /batches/:id/files (no SAS)
      - /pipeline/*                    -> /classifications/*
      - /hitl/queue                    -> /classifications/review
      - POST /hitl/queue/:id/review    -> PATCH /classifications/review/:id

    Internal code/DB now use `batches` / `hitl_queue` (renamed from
    `declaration_runs` in migration 0084). The API surface continues to
    use `/batches/*`. See tracker/API_AUDIT_SPEC_2026-05-12.md.

    Authentication: every endpoint outside `/health` requires BOTH:
      1. A valid Microsoft Entra ID access token, validated by APIM's
         `validate-jwt` inbound policy. Audience must be the API app
         registration's AppId URI (or the app's client_id GUID as
         alternate audience). 401 if missing/invalid/expired.
      2. APIM gateway shared-secret header (`x-apim-shared-secret`),
...
         named value `apim-shared-secret`. APIM strips any
...

...

...

...
servers:
...
components:
  securitySchemes:
    apimSharedSecret:
      type: apiKey
      in: header
      name: x-apim-shared-secret
      description: |
        Shared secret injected by APIM's inbound policy. Not a user-facing
        credential; users authenticate via Entra (see `entraJwt`).
    entraJwt:
      type: oauth2
      description: |
        Microsoft Entra ID user token. Validated by APIM's `validate-jwt`
        policy before the request reaches the backend.
      flows:
        authorizationCode:
          authorizationUrl: https://login.microsoftonline.com/{tenant}/oauth2/v2.0/authorize
          tokenUrl: https://login.microsoftonline.com/{tenant}/oauth2/v2.0/token
          scopes:
            api://clearai-backend/.default: Default access scope
  parameters:
    BatchId:
      name: id
      in: path
      required: true
      description: UUIDv7 of a batch.
      schema:
        type: string
        format: uuid
...
  schemas:
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              description: Machine-readable error code (snake_case).
              example: batch_validation_failed
            message:
              type: string
              description: Human-readable explanation.
            details:
              type: object
              additionalProperties: true
              description: Optional structured detail (Zod issues, field name, etc.).

    # ---------- probes ----------
...
    BatchMode:
      type: string
      enum: [classify_only, classify_and_declare]
      description: |
        `classify_only` runs Phase 1 and stops; the batch produces no XML.
        `classify_and_declare` runs Phase 1 then Phase 2 (rendered XML
        per HV item / per LV bundle). Default `classify_and_declare`.
    BatchStatus:
      type: string
      enum: [pending, running, completed, failed, cancelled]
    ClassificationPhase:
      type: string
      enum: [pending, running, completed, failed]
    DeclarationStatus:
      type: string
      nullable: true
      enum: [pending, running, completed, failed, skipped, null]
...
    BatchCreated:
      type: object
      required: [batch_id, mode]
      properties:
        batch_id:
          type: string
          format: uuid
        mode:
          $ref: '#/components/schemas/BatchMode'
    BatchSummary:
      type: object
      required:
        [id, operator_slug, mode, status, classification_status, declaration_status,
         row_count, succeeded, flagged, blocked, failed, pending, pending_infra,
         started_at, completed_at, error]
      properties:
        id:
          type: string
          format: uuid
        operator_slug:
          type: string
          description: |
            The operator that owns this batch (e.g. "naqel"). Read-only;
            single-operator V1, always "naqel".
        mode:
          $ref: '#/components/schemas/BatchMode'
        status:
          $ref: '#/components/schemas/BatchStatus'
        classification_status:
          $ref: '#/components/schemas/ClassificationPhase'
        declaration_status:
          $ref: '#/components/schemas/DeclarationStatus'
        row_count:
          type: integer
          description: Total parsed rows in this batch.
        succeeded: { type: integer }
        flagged: { type: integer }
        blocked: { type: integer }
        failed: { type: integer }
        pending: { type: integer }
        pending_infra:
          type: integer
          description: |
            Rows downgraded by an LLM-stage exhaustion (graceful_degrade)
            rather than a real bad-data result. Surfaced separately so the
            HITL UI can hide infra-only failures from the reviewer queue.
        started_at:
          type: string
          nullable: true
          format: date-time
        completed_at:
          type: string
          nullable: true
          format: date-time
        error:
          type: string
          nullable: true
...
    DeclaredValue:
      type: object
      required: [hs_code, description, amount, currency]
      description: |
        Merchant-supplied values exactly as they appeared in the source row,
        before any FX conversion or pipeline cleanup.
      properties:
        hs_code:
          type: string
          nullable: true
          description: Merchant's HS code prefix as supplied. Null when absent or malformed.
        description:
          type: string
          nullable: true
          description: Verbatim merchant description.
        amount:
          type: number
          nullable: true
          description: Declared value in the merchant's submitted currency.
        currency:
          type: string
          nullable: true
          pattern: '^[A-Z]{3}$'
          description: ISO 4217 currency code as supplied by the merchant.
    ResolvedHsCodeDescription:
      type: object
      required: [full_hierarchy, zatca_submission_description, retrieval_query]
      properties:
        full_hierarchy:
          type: array
          minItems: 2
          maxItems: 2
          items:
            $ref: '#/components/schemas/LocalizedString'
          description: Bilingual catalog breadcrumb (en + ar) for the resolved HS code.
        zatca_submission_description:
          type: array
          minItems: 2
          maxItems: 2
          items:
            $ref: '#/components/schemas/LocalizedString'
          description: |
            LLM-generated Arabic goods description that ships in the ZATCA
            envelope. English row is null until the bilingual surface lands.
        retrieval_query:
          type: string
          nullable: true
          description: |
            Effective description used by Track A retrieval (post-cleanup /
            researcher). Lets operators see what the pipeline actually queried
            against vs. the verbatim merchant input.
...
    Value:
      type: object
      required: [amount, rate, rate_as_of]
      description: |
        SAR-denominated declared value (ZATCA envelopes are SAR-only) plus
        the FX rate + as-of date used at parse time.
      properties:
        amount:
          $ref: '#/components/schemas/CanonicalValueAmount'
        rate:
          type: number
          nullable: true
          description: SAR-per-unit rate applied to the merchant's value (1 for SAR itself).
        rate_as_of:
          type: string
          nullable: true
          format: date
          description: Calendar date of the fx_rates row used for conversion (Asia/Riyadh).
    ClassificationResult:
      type: object
      required: [resolved_hs_code, classification_status, classification_confidence, sanity_verdict]
      properties:
        resolved_hs_code:
          type: string
          nullable: true
          pattern: '^\d{12}$'
          description: 12-digit code chosen by the pipeline. Null when no defensible answer.
        classification_status:
          type: string
          nullable: true
          enum: [AGREEMENT, DRIFT, ZERO_SIGNAL]
          description: V1 reconciliation outcome between the two tracks.
        classification_confidence:
          type: number
          nullable: true
          minimum: 0
          maximum: 100
          description: 0-100 score. Always null in V1 until the score work lands.
        sanity_verdict:
          type: string
          nullable: true
          enum: [PASS, FLAG, BLOCK]
          description: Stage 3 value-plausibility verdict.
    CanonicalItem:
      type: object
      required:
        - id
        - declared_value
        - resolved_hs_code_description
        - value
        - duty_info
        - procedures
        - classification_result
        - error
      description: |
        Unified per-item shape returned by /batches/{id}/items (batch listing),
        /classifications/dispatch (single-shot), and /classifications/{id}
        (single-item lookup). `row_index` is present only on batch responses;
        `trace` is present only when `?include_trace=true`. `operator_slug`
        is NOT on the item — it lives on the envelope (batch level for batch
        endpoints, response root for standalone).
      properties:
        id:
          type: string
          format: uuid
        row_index:
          type: integer
          description: 1-based row position. Batch listing only; omitted on standalone.
        declared_value:
          $ref: '#/components/schemas/DeclaredValue'
        resolved_hs_code_description:
          $ref: '#/components/schemas/ResolvedHsCodeDescription'
        value:
          $ref: '#/components/schemas/Value'
        duty_info:
          $ref: '#/components/schemas/DutyInfo'
          nullable: true
        procedures:
          type: array
          items:
            $ref: '#/components/schemas/ProcedureInfo'
        classification_result:
          $ref: '#/components/schemas/ClassificationResult'
        trace:
          type: object
          additionalProperties: true
          description: |
            Full PipelineTrace. Present only when the request included
            `?include_trace=true`; otherwise the key is omitted entirely.
        error:
          type: string
          nullable: true
    DutyInfo:
      type: object
      required: [rate_percent, status]
      description: |
        Duty information for an HS code. Exactly one of `rate_percent` / `status`
        carries data; the other is null. Returned as null at the parent level
        when the code has no duty data.
      properties:
        rate_percent:
          type: number
          nullable: true
          description: Numeric % when the row carries a rate; null when the row is a non-rate status.
        status:
          type: string
          nullable: true
          enum: [exempted, prohibited_import, prohibited_export, prohibited_both]
          description: Non-rate duty status; null when the row carries a rate.
    ProcedureInfo:
      type: object
      required: [code, description_ar, is_repealed]
      properties:
        code:
          type: string
        description_ar:
          type: string
        is_repealed:
          type: boolean
          description: True when the procedure description ends with `(ملغي)`.
    BatchItemsPage:
      type: object
      required: [batch_id, operator_slug, classification_phase, total, limit, offset, has_more, next_offset, items]
      properties:
        batch_id:
          type: string
          format: uuid
        operator_slug:
          type: string
          example: naqel
          description: |
            Operator the batch belongs to. Surfaced at envelope level
            rather than per item — applies uniformly to every row.
        classification_phase:
          $ref: '#/components/schemas/ClassificationPhase'
        total:
          type: integer
          description: Total items in the batch (across all pages).
        limit:
          type: integer
        offset:
          type: integer
        has_more:
          type: boolean
        next_offset:
          type: integer
          nullable: true
          description: Offset to pass on the next request, or null when has_more=false.
        items:
          type: array
          items:
            $ref: '#/components/schemas/CanonicalItem'
    BatchFile:
      type: object
      required: [name]
      properties:
        name:
          type: string
          example: hv/019e15e7-3e97-764f-845e-8848d7753193.xml
        size_bytes:
          type: integer
          nullable: true
        content_type:
          type: string
          nullable: true
    BatchFilesList:
      type: object
      required: [batch_id, files]
      properties:
        batch_id:
          type: string
          format: uuid
        files:
          type: array
          items:
            $ref: '#/components/schemas/BatchFile'

    # ---------- submission-description ----------
...
paths:
...
  /batches:
    get:
      operationId: list-batches
      tags: [batches]
      summary: List batches with optional filters
      description: |
        Paginated list of declaration runs (batches), newest first. Returns
        slim BatchSummary rows (no per-item trace). Use `GET /batches/{id}`
        for a single run or `GET /batches/{id}/items` for its items.

        `status` accepts a comma-separated list. Valid values are the
        BatchStatus enum: `pending`, `running`, `completed`, `failed`,
        `cancelled`. Note: there is no `processing` status — use `running`.

        `created_since` / `created_until` accept ISO-8601 timestamps and
        bound the inclusive `created_at` window.
      security:
        - apimSharedSecret: []
          entraJwt: ['api://clearai-backend/.default']
      parameters:
        - name: limit
          in: query
          required: false
          schema: { type: integer, minimum: 1, maximum: 500, default: 50 }
        - name: offset
          in: query
          required: false
          schema: { type: integer, minimum: 0, default: 0 }
        - name: status
          in: query
          required: false
          description: Comma-separated subset of BatchStatus values.
          schema: { type: string, example: 'running,completed' }
        - name: created_since
          in: query
          required: false
          schema: { type: string, format: date-time }
        - name: created_until
          in: query
          required: false
          schema: { type: string, format: date-time }
      responses:
        '200':
          description: Paginated list of batches.
          content:
            application/json:
              schema:
                type: object
                required: [items, total, limit, offset, has_more, next_offset]
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/BatchSummary'
                  total: { type: integer }
                  limit: { type: integer }
                  offset: { type: integer }
                  has_more: { type: boolean }
                  next_offset:
                    type: integer
                    nullable: true
        '400':
          description: Invalid query (bad status value, bad date, bad pagination).
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '401':
          description: Origin access denied.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }

    post:
      operationId: create-batch
      tags: [batches]
      summary: Upload an xlsx/csv and start a batch
      description: |
        Multipart upload. The server validates the file extension, parses
        rows synchronously into the canonical shape, persists the batch, and
        kicks off Phase 1 (classification) in the background. Returns 202
        immediately with the batch id.

        2026-05-12 cutover: `operator_slug`, `callback_url`, `metadata`
        removed from the multipart body. Single-operator V1 ("naqel"
        hardcoded server-side).
      security:
        - apimSharedSecret: []
          entraJwt: ['api://clearai-backend/.default']
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
                  description: '.csv or .xlsx commercial-invoice file. Max 25 MB.'
                mode:
                  $ref: '#/components/schemas/BatchMode'
      responses:
        '202':
          description: Batch accepted; processing in background.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BatchCreated'
        '400':
          description: Invalid body (bad file type, bad mode, etc.).
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '413':
          description: |
            Upload exceeds row count limit (batch_too_large) or 25 MB file size limit.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '401':
          description: Origin access denied (missing/invalid APIM secret in production).
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /batches/{id}:
    parameters:
      - $ref: '#/components/parameters/BatchId'

    get:
      operationId: get-batch
      tags: [batches]
      summary: Batch summary
      security:
        - apimSharedSecret: []
          entraJwt: ['api://clearai-backend/.default']
      responses:
        '200':
          description: Batch found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BatchSummary'
        '404':
          description: No batch with that id.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '401':
          description: Origin access denied.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /batches/{id}/items:
    parameters:
      - $ref: '#/components/parameters/BatchId'

    get:
      operationId: get-batch-items
      tags: [batches]
      summary: Per-item classification results (paged)
      description: |
        Page of items + classification result + trace. 2026-05-12 cutover:
        added `total` / `has_more` / `next_offset` for ergonomic paging.
        Envelope key is `batch_id`.
      security:
        - apimSharedSecret: []
          entraJwt: ['api://clearai-backend/.default']
      parameters:
        - name: limit
          in: query
          required: false
          schema: { type: integer, minimum: 1, maximum: 500, default: 100 }
        - name: offset
          in: query
          required: false
          schema: { type: integer, minimum: 0, default: 0 }
        - name: include_trace
          in: query
          required: false
          description: When true, include the full per-item `trace` object. Off by default to keep paging cheap.
          schema: { type: boolean, default: false }
      responses:
        '200':
          description: Page of items.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BatchItemsPage'
        '425':
          description: Phase 1 still pending or running.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '404':
          description: No batch with that id.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /batches/{id}/cancel:
    parameters:
      - $ref: '#/components/parameters/BatchId'

    post:
      operationId: cancel-batch
      tags: [batches]
      summary: Cancel a running batch
      description: |
        Replaces the old `PATCH /declaration-runs/{id}` cancel. No body.
        Idempotent on already-cancelled. Returns 409 on already-terminal
        non-cancelled states (completed/failed).
      security:
        - apimSharedSecret: []
          entraJwt: ['api://clearai-backend/.default']
      responses:
        '200':
          description: Cancelled.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BatchSummary'
        '404':
          description: No batch with that id.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '409':
          description: Batch already in a terminal state.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /batches/{id}/files:
    parameters:
      - $ref: '#/components/parameters/BatchId'

    get:
      operationId: get-batch-files
      tags: [batches]
      summary: List files produced by the batch
      description: |
        Replaces the old `/download-links` endpoint. No SAS URLs minted —
        the SPA streams each file via `/batches/{id}/files/*`. Response
        keys are snake_case (`size_bytes`, `content_type`).

        Returns 409 `no_blob_prefix` when the batch hasn't produced any
        XML output (legacy or in-flight).
      security:
        - apimSharedSecret: []
          entraJwt: ['api://clearai-backend/.default']
      responses:
        '200':
          description: List of files under the batch's blob prefix.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BatchFilesList'
        '404':
          description: No batch with that id, or no files under its prefix.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '409':
          description: Batch has no blob output yet.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }

  # APIM routing note: `urlTemplate` matches a SINGLE path segment per
  # `{param}` placeholder. Blob paths like `lv/<filingId>.xml` contain a
  # slash and cannot match a single template parameter, so the flat
  # `/files/{path}` operation 404s at the gateway for nested paths.
  # Backend (Fastify) uses a `*` wildcard route — `/batches/:id/files/*` —
  # which accepts any sub-path. We split into THREE explicit operations:
  # one flat (input.csv, run-index.json, classifications.json) and one
  # each for `hv/` and `lv/` nested files. All three forward to the same
  # backend route. If a new sub-folder is added (e.g. `errors/`), add a
  # matching operation here AND re-import.
  /batches/{id}/files/{path}:
    parameters:
      - $ref: '#/components/parameters/BatchId'
      - name: path
        in: path
        required: true
        description: |
          Flat file name (no slashes), one of `input.csv`, `input.xlsx`,
          `run-index.json`, `classifications.json`. For nested files
          under `hv/` or `lv/`, use the dedicated operations below.
        schema: { type: string, example: run-index.json }

    get:
      operationId: get-batch-file-flat
      tags: [batches]
      summary: Stream a flat file from a batch
      description: |
        Streams bytes through the backend. Path-traversal guard rejects
        `..`, leading `/`, and `\\`.
      security:
        - apimSharedSecret: []
          entraJwt: ['api://clearai-backend/.default']
      responses:
        '200':
          description: File bytes (XML, JSON, CSV, or octet-stream).
          content:
            '*/*': {}
        '400':
          description: Invalid path or id.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '404':
          description: No batch with that id, or no file at that path.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
...
security:
  - apimSharedSecret: []
    entraJwt: ['api://clearai-backend/.default']